pyproject.toml
[project]
name = "ai-chat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "arq>=0.25.0",
    "asyncpg>=0.31.0",
    "clerk-backend-api>=5.0.6",
    "fastapi>=0.135.3",
    "fastapi-limiter==0.1.6",
    "groq>=1.1.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "razorpay>=2.0.1",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
